// apps/client/src/app/probable-waffle/game/prefabs/characters/skaduwee/skaduwee-worker/skaduwee-worker-male/SkaduweeWorkerMaleSfx.ts
export enum SkaduweeWorkerMaleSfx {
  ATTACK_1 = "Attack 1",
  ATTACK_2 = "Attack 2",
  ATTACK_3 = "Attack 3",
  BUILD_1 = "Build 1",
  BUILD_2 = "Build 2",
  CHOP_1 = "Chop 1",
  CHOP_2 = "Chop 2",
  DAMAGE_1 = "Damage 1",
  DAMAGE_2 = "Damage 2",
  DAMAGE_3 = "Damage 3",
  DEATH_1 = "Death 1",
  DEATH_2 = "Death 2",
  DEATH_3 = "Death 3",
  ENTER_BUILDING_1 = "Enter building 1",
  ENTER_BUILDING_2 = "Enter building 2",
  LOCATION_UNAVAILABLE_1 = "Location unavailable 1",
  LOCATION_UNAVAILABLE_2 = "Location unavailable 2",
  MINE_1 = "Mine 1",
  MINE_2 = "Mine 2",
  MOVE_1 = "Move 1",
  MOVE_2 = "Move 2",
  MOVE_3 = "Move 3",
  MOVE_4 = "Move 4",
  REPAIR_1 = "Repair 1",
  REPAIR_2 = "Repair 2",
  SELECTION_1 = "Selection 1",
  SELECTION_2 = "Selection 2",
  SELECTION_3 = "Selection 3",
  SELECTION_4 = "Selection 4"
}

export const SkaduweeWorkerMaleSfxAttackSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.ATTACK_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.ATTACK_2 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.ATTACK_3 }
];

export const SkaduweeWorkerMaleSfxBuildSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.BUILD_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.BUILD_2 }
];

export const SkaduweeWorkerMaleSfxChopSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.CHOP_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.CHOP_2 }
];

export const SkaduweeWorkerMaleSfxDamageSounds: SoundDefinition[] = [
  // #366 - damage sound effects sound "wrong" during combat
  // { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.DAMAGE_1 },
  // { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.DAMAGE_2 },
  // { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.DAMAGE_3 }
];

export const SkaduweeWorkerMaleSfxDeathSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.DEATH_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.DEATH_2 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.DEATH_3 }
];

export const SkaduweeWorkerMaleSfxEnterSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.ENTER_BUILDING_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.ENTER_BUILDING_2 }
];

export const SkaduweeWorkerMaleSfxLocationSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.LOCATION_UNAVAILABLE_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.LOCATION_UNAVAILABLE_2 }
];

export const SkaduweeWorkerMaleSfxMineSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.MINE_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.MINE_2 }
];

export const SkaduweeWorkerMaleSfxMoveSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.MOVE_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.MOVE_2 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.MOVE_3 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.MOVE_4 }
];

export const SkaduweeWorkerMaleSfxRepairSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.REPAIR_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.REPAIR_2 }
];

export const SkaduweeWorkerMaleSfxSelectionSounds: SoundDefinition[] = [
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.SELECTION_1 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.SELECTION_2 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.SELECTION_3 },
  { key: AudioSprites.SKADUWEE_WORKER_MALE, spriteName: SkaduweeWorkerMaleSfx.SELECTION_4 }
];